HW: First version of PCB layout
First version of PCB layout is ready for review.
Changed the assignments of pins for switches and status LEDs
to ease the routing of traces.
PCB's bottom layer is GND plane, most of the traces are placed
on top layer. Ground planes are connected with via stitching on
full area of the PCB.
There are no high speed signals on this PCB that might be of a concern
while routing so general rules of keeping the traces short and diff pairs
of the same lenghts are used.

Signed-off-by: AlexeyZ <silkt207@gmail.com>